sem5/SiMZIIS/lab5/src/pyproject.toml
[tool.poetry]
name = "lab5"
version = "0.1.0"
description = ""
authors = ["Ehor Hleoza <gleoza04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pycryptodome = "^3.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
